# back-end-challenge/back-end-challenge/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "back_end_challenge",
    "version": "v1"
  },
  "paths": {
...
    "/Timeline/AddComment": {
      "post": {
        "tags": [
          "Timeline"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TimelineDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/Timeline/GetTimeline": {
      "get": {
        "tags": [
          "Timeline"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "user",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TimelineDto"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "FollowingDto": {
      "type": "object",
      "properties": {
        "idFollowing": {
          "format": "int32",
          "type": "integer"
        },
        "userMaster": {
          "type": "string"
        },
        "userFollowing": {
          "type": "string"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "userMasterNavigation": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "TimelineDto": {
      "type": "object",
      "properties": {
        "idTimeline": {
          "format": "int32",
          "type": "integer"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "user": {
          "type": "string"
        },
        "userNavigation": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "UserDto": {
      "type": "object",
      "properties": {
        "idUser": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "activate": {
          "format": "int32",
          "type": "integer"
        },
        "thought": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "followers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FollowerDto"
          }
        },
        "followings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FollowingDto"
          }
        },
        "timelines": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TimelineDto"
          }
        }
      }
    },
    "FollowerDto": {
      "type": "object",
      "properties": {
        "idFollower": {
          "format": "int32",
          "type": "integer"
        },
        "userMaster": {
          "type": "string"
        },
        "userFollower": {
          "type": "string"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "userMasterNavigation": {
          "$ref": "#/definitions/UserDto"
        }
      }
    }
  }
}